Add CurlHeaderProperty component and delete CurlPropertyInput and include it CurlProperties
